added some ssao code that does nothing

+++ includes/opengl/Texture2D.h
@@ -7,7 +7,7 @@ class Texture2D
 friend class FrameBuffer;
 public:
     Texture2D();
-    Texture2D(const glm::ivec2 size, GLuint format = GL_RGBA);
+    Texture2D(const glm::uvec2 size, GLuint format = GL_RGBA);
     ~Texture2D();
 
     void Bind() const;
@@ -15,12 +15,12 @@ public:
 
     void SetInternalFormat(GLuint format);
 
-    void Resize(glm::ivec2 size);
-    void SetData(glm::vec2 pos, glm::vec2 size, unsigned char* data);
-    glm::vec2 getSize() {return m_Size;}
+    void Resize(glm::uvec2 size);
+    void SetData(glm::vec2 pos, glm::uvec2 size, unsigned char* data);
+    glm::uvec2 getSize() {return m_Size;}
 
 private:
     GLuint m_RendererID;
     GLuint m_InternalFormat;
-    glm::vec2 m_Size;
+    glm::uvec2 m_Size;
 };
